# backend/openapi.yaml
openapi: 3.0.1
info:
  title: AguaVigia CTG — API publica
  description: |-
    Monitoreo ciudadano del acueducto de Cartagena de Indias.

    Plataforma ciudadana e independiente: no esta afiliada a Aguas de Cartagena
    S.A. E.S.P. ni a ninguna entidad distrital.
  contact:
    name: Equipo AguaVigia CTG — Tecnologico Comfenalco
  license:
    name: Proyecto de aula
  version: 1.0.0
servers:
- url: /
  description: Mismo origen que sirve la API
...
paths:
...
  /api/veedor/sesion:
    post:
      tags:
      - Veedor
      summary: Iniciar sesion en el panel del veedor
      description: |-
        Devuelve un token JWT valido por 8 horas (RNF011) junto con el rol y los permisos
        ya resueltos. Si la cuenta tiene segundo factor y no se envio `codigoTotp`, la
        respuesta es 401 con type `segundo-factor-requerido`: hay que reintentar con el
        codigo, no es un error de credencial.
      operationId: iniciarSesion
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CredencialVeedor"
        required: true
      responses:
        "200":
          description: "Credencial correcta, token emitido"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SesionVeedor"
        "401":
          description: "Credencial incorrecta, o falta el segundo factor"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SesionVeedor"
        "403":
          description: La cuenta existe pero no esta habilitada para entrar
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SesionVeedor"
        "423":
          description: Cuenta bloqueada por intentos fallidos
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SesionVeedor"
        "429":
          description: Demasiados intentos desde esta IP
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SesionVeedor"
...
components:
  schemas:
...
    CredencialVeedor:
      type: object
      properties:
        correo:
          minLength: 1
          type: string
          description: Correo de la cuenta
          example: veedor@ejemplo.org
        clave:
          minLength: 1
          type: string
          description: Clave de la cuenta
        codigoTotp:
          type: string
          description: |-
            Codigo de 6 digitos de la app de autenticacion. Se omite en el primer intento; si la
            cuenta tiene segundo factor, la respuesta 401 con type `segundo-factor-requerido`
            indica que hay que reintentar incluyendolo.
      description: Credencial de acceso al panel del veedor
    SesionVeedor:
      type: object
      properties:
        token:
          type: string
          description: "Token JWT. Se envia como 'Authorization: Bearer <token>'"
        usuarioId:
          type: string
        nombre:
          type: string
        correo:
          type: string
        rol:
          type: string
          description: "ADMIN, VEEDOR u OBSERVADOR"
        permisos:
          type: array
          description: "Permisos efectivos ya resueltos: rol mas concedidos menos\
            \ revocados"
          items:
            type: string
            description: "Permisos efectivos ya resueltos: rol mas concedidos menos\
              \ revocados"
        alcance:
          type: string
          description: |-
            COMPLETO, o ALTA_SEGUNDO_FACTOR cuando la cuenta es ADMIN y todavia no dio de alta su
            TOTP. Con ese alcance el token solo sirve para /api/veedor/segundo-factor.
      description: "Sesion emitida para el panel del veedor (RNF011: expira en 8 horas)"
...